feat(clipboard): enhance clipboard functionality with content serialization
- Refactored `ClipboardController` to utilize a new `readSelectedContent` method from `DomModel`, simplifying the handling of copied content.
- Introduced `serializeRange` method in `ParseController` to streamline the serialization of selected text tokens.
- Added a new utility function `serializeRange` in `utils/serializeRange.ts` for better token management during clipboard operations.
- Removed deprecated methods related to range handling in `ClipboardController` for cleaner code.

## Test Plan
- [x] All existing tests pass
- [ ] Manual testing of clipboard operations in the editor
